COMPLETE Dashboard Display
I have finished the dashboard display to show all the differnt
options that are specified by the diagram on the assignment.

However, there is a bug that causes the writing to become harder
to read if there is more than so many characters on one line.
We should try to fix this if we have time.